core/remote: Move `Warning` type definition
  The `Warning` type was defined in its own file with nothing else,
  adding yet another file to the code base for a questionable benefit.
  Since this type is tighlty tied to the Cozy we decided to move its
  definition with the `RemoteCozy` class.
