--- core/remote/warning.js
@@ -1,15 +1,0 @@
-/**
- * @module core/remote/warning
- * @flow
- */
-
-/*::
-export type Warning = {
-  error: string,
-  title: string,
-  detail: string,
-  links: {
-    self: string
-  }
-}
-*/
